refactor final v.1
